fix: folder rename to match original structure
