+++ packages/elements-core/src/components/TableOfContents/types.ts
@@ -1,22 +1,5 @@
-import type { TableOfContentsItem } from '@stoplight/elements-utils';
-
-export type TableOfContentsProps = {
-  tree: TableOfContentsItem[];
-  activeId: string;
-  Link: CustomLinkComponent;
-  maxDepthOpenByDefault?: number;
-  externalScrollbar?: boolean;
-  isInResponsiveMode?: boolean;
-  onLinkClick?(): void;
-};
-
-export type CustomLinkComponent = React.ComponentType<{
-  to: string;
-  className?: string;
-  children: React.ReactNode;
-}>;
-
 export type {
+  CustomLinkComponent,
   TableOfContentsDivider,
   TableOfContentsExternalLink,
   TableOfContentsGroup,
@@ -24,4 +7,5 @@ export type {
   TableOfContentsItem,
   TableOfContentsNode,
   TableOfContentsNodeGroup,
+  TableOfContentsProps,
 } from '@stoplight/elements-utils';
